+++ app/lib/iiif_manifest_factory.rb
@@ -1,25 +1,69 @@
 # frozen_string_literal: true
 
 class IiifManifestFactory
-  attr_reader :oid, :manifest, :parent_object
+  attr_reader :oid, :manifest, :parent_object, :mets_doc
 
   require 'iiif/presentation'
 
   def initialize(oid)
+    @manifest_base_url = ENV["IIIF_MANIFESTS_BASE_URL"] || "http://localhost/manifests/"
+    @image_base_url = ENV["IIIF_IMAGE_BASE_URL"] || "http://localhost:8182/iiif"
     @oid = oid
     @parent_object = ParentObject.find_by(oid: oid)
+    mets_object = MetsXmlImport.find_by(oid: oid)
+    iiif_info = { oid: oid, reason: "No METs xml available for #{oid}" }
+    raise ActiveRecord::RecordNotFound, "Unable to create iiif manifest for: #{iiif_info.to_json}" unless mets_object
+    @mets_doc = MetsDocument.new(mets_object.mets_xml)
     @manifest = construct_manifest
   end
 
   # Build the actual manifest object
   def construct_manifest
     manifest = IIIF::Presentation::Manifest.new(seed)
+    manifest.sequences << create_first_sequence(oid)
+    add_canvases_to_sequence(manifest.sequences.first)
     manifest
+  end
+
+  def create_first_sequence(oid)
+    sequence = IIIF::Presentation::Sequence.new
+    sequence["@id"] = "#{ENV['IIIF_MANIFESTS_BASE_URL']}/sequence/#{oid}"
+    sequence
+  end
+
+  def add_image_to_canvas(file, canvas)
+    images = canvas.images
+    image = IIIF::Presentation::Resource.new
+    image['@id'] = "#{@manifest_base_url}/oid/#{@oid}/canvas/#{file[:image_id]}/image/1"
+    image['@type'] = "oa:Annotation"
+    image["motivation"] = "sc:painting"
+    image["on"] = "#{@manifest_base_url}/oid/#{@oid}/canvas/#{file[:image_id]}"
+
+    image_resource = IIIF::Presentation::ImageResource.create_image_api_image_resource(
+      service_id: "#{@image_base_url}/2/#{file[:image_id]}"
+    )
+    image["resource"] = image_resource
+    images << image
+  end
+
+  def add_canvases_to_sequence(sequence)
+    canvases = sequence.canvases
+    image_files = @mets_doc.combined
+    image_files.sort_by! { |file| file[:order].to_i }
+    image_files.map do |file|
+      canvas = IIIF::Presentation::Canvas.new
+      canvas['@id'] = "#{@manifest_base_url}/oid/#{@oid}/canvas/#{file[:image_id]}"
+      canvas['label'] = file[:order_label]
+      add_image_to_canvas(file, canvas)
+      canvas['height'] = canvas.images.first["resource"]["height"]
+      canvas['width'] = canvas.images.first["resource"]["width"]
+      canvases << canvas
+    end
   end
 
   def seed
     {
-      '@id' => "http://127.0.0.1/manifests/#{@oid}.json",
+      '@id' => "#{ENV['IIIF_MANIFESTS_BASE_URL']}/#{@oid}.json",
       'label' => @parent_object.authoritative_json["title"].first
     }
   end