+++ app/models/child_object.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class ChildObject < ApplicationRecord
+  belongs_to :parent_object, foreign_key: 'parent_object_oid', class_name: "ParentObject"
+  self.primary_key = 'child_oid'
+end