+++ db/migrate/20200813124248_parent_oid_to_integer.rb
@@ -1,0 +1,6 @@
+class ParentOidToInteger < ActiveRecord::Migration[6.0]
+  def change
+    change_column :parent_objects, :oid, 'bigint USING CAST(oid AS bigint)', primary_key: true
+    remove_column :parent_objects, :id
+  end
+end