+++ db/migrate/20200816184926_add_child_object_count_to_parent_objects.rb
@@ -1,0 +1,5 @@
+class AddChildObjectCountToParentObjects < ActiveRecord::Migration[6.0]
+  def change
+    add_column :parent_objects, :child_object_count, :integer
+  end
+end