+++ spec/factories/goobi_xml_imports.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-FactoryBot.define do
-  factory :goobi_xml_import do
-    goobi_xml { "" }
-  end
-end