+++ spec/factories/mets_xml_imports.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :mets_xml_import do
+    mets_xml { "" }
+    oid { "2012315" }
+  end
+end