+++ spec/models/concerns/json_file_spec.rb
@@ -2,11 +2,10 @@
 require "rails_helper"
 
 RSpec.describe JsonFile, prep_metadata_sources: true do
-  let(:parent_object) { FactoryBot.create(:parent_object, oid: '100001') }
-  let(:path_to_example_file) { Rails.root.join("spec", "fixtures", "ladybird", "100001.json") }
+  let(:parent_object) { FactoryBot.create(:parent_object, oid: '2003431') }
+  let(:path_to_example_file) { Rails.root.join("spec", "fixtures", "ladybird", "2003431.json") }
   before do
-    stub_request(:get, "https://yul-development-samples.s3.amazonaws.com/ladybird/100001.json")
-      .to_return(status: 200, body: File.open(File.join(fixture_path, "ladybird", "100001.json")).read)
+    stub_metadata_cloud("2003431")
   end
 
   it "can save a ParentObject to json" do