+++ spec/rails_helper.rb
@@ -34,7 +34,7 @@ rescue ActiveRecord::PendingMigrationError => e
 end
 RSpec.configure do |config|
   config.include(AuthHelper, type: :request)
-  config.include(MetdataCallHelper)
+  config.include(MetdataSourcesHelper)
   config.include(SolrHelper)
   config.include(StubRequestHelper)
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
@@ -44,13 +44,6 @@ RSpec.configure do |config|
   # examples within a transaction, remove the following line or assign false
   # instead of true.
   config.use_transactional_fixtures = true
-
-  # Clear out solr before test
-  config.before(clean: true) do
-    solr = SolrService.connection
-    solr.delete_by_query '*:*'
-    solr.commit
-  end
 
   # You can uncomment this line to turn off ActiveRecord support entirely.
   # config.use_active_record = false