+++ spec/support/stub_requests_helper.rb
@@ -2,7 +2,14 @@
 
 module StubRequestHelper
   def stub_metadata_cloud(oid, source_name = 'ladybird')
-    stub_request(:get, "https://yul-development-samples.s3.amazonaws.com/#{source_name}/#{oid}.json")
-      .to_return(status: 200, body: File.open(File.join(fixture_path, source_name, "#{oid}.json")))
+    vpn = ENV["VPN"]
+    allowed_sites = ["solr", MetadataCloudService.metadata_cloud_host]
+    if vpn == "true"
+      WebMock.disable_net_connect!(allow: allowed_sites)
+      stub_request(:put, "https://#{ENV['SAMPLE_BUCKET']}.s3.amazonaws.com/#{source_name}/#{oid}.json").to_return(status: 200)
+    else
+      stub_request(:get, "https://#{ENV['SAMPLE_BUCKET']}.s3.amazonaws.com/#{source_name}/#{oid}.json")
+        .to_return(status: 200, body: File.open(File.join(fixture_path, source_name, "#{oid}.json")))
+    end
   end
 end